feat: move google sign in to the top
Will make it more prominent.
