+++ src/lib/text/app.document.json
@@ -22,5 +22,6 @@
   "favorites.empty": "You have no favorites!",
   "navigation.login.google": "Login with Google",
   "navigation.register.google": "Register with Google",
-  "navigation.contact": "Contact"
+  "navigation.contact": "Contact",
+  "navigation.login.title": "Welcome!"
 }
