+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '../../components/forms/LoginForm';
+import LoginForm from './components/LoginForm';
 import { Container } from '../../components/styled';
